Move `_speed_to_lsb()` into the `EMC2101` base class as a method.
This conversion is going to depend on the PWM_F and DAC register values, so it needs
to be a method. Use this method in `manual_fan_speed()` as well as the LUT subclass.

Move the `_pwm_freq` register definition into the base class too. It will be
needed for the fixed fan speed conversions.

Also import the `MAX_LUT_*` constants from the base module rather than redefining them.